// src/types.ts
export interface ClockSettings {
  background: string;
  backgroundImage: string;
  clockSize: string;
  clockColor: string;
  fontFamily: string;
  showSeconds: boolean;
  showDate: boolean;
  orientation: 'horizontal' | 'vertical';
  useCardsMode: boolean;
  is24Hour: boolean;
  showPomodoro: boolean;
  isVertical: boolean;
  useDepthEffect: boolean;
  selectedFont: string;
  clockFace: string;
} 